eval/requirements.txt
arize>=7.19.0
claude-agent-sdk>=0.1.0
anthropic
python-dotenv
